% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{adjusted_rand}
\alias{adjusted_rand}
\title{A "soft" \emph{and adjusted} version of a rand index between two lists of responsibility
(membership probability) matrices. Measures the /disagreement/ between the two clusterings.}
\usage{
adjusted_rand(resp_list1, resp_list2, times = NULL)
}
\arguments{
\item{resp_list1}{One list of responsibility matrices.}

\item{resp_list2}{Another list of responsibility matrices.}

\item{times}{Optional; if you would like to isolate your attention to some
specific times.}
}
\value{
A single soft rand index for all particles across all times.
}
\description{
A "soft" \emph{and adjusted} version of a rand index between two lists of responsibility
(membership probability) matrices. Measures the /disagreement/ between the two clusterings.
}