% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{adjusted_rand_onetime}
\alias{adjusted_rand_onetime}
\title{A "soft" \emph{and adjusted} version of a rand index between two responsibility
(membership probability) matrices. Measures the /disagreement/ between the
two clusterings.}
\usage{
adjusted_rand_onetime(resp1, resp2)
}
\arguments{
\item{resp1}{One responsibility matrices.}

\item{resp2}{Another responsibility matrices.}
}
\value{
A single soft rand index for all particles.
}
\description{
A "soft" \emph{and adjusted} version of a rand index between two responsibility
(membership probability) matrices. Measures the /disagreement/ between the
two clusterings.
}