% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{admm_oneclust}
\alias{admm_oneclust}
\title{One cluster's admm for a fixed step size (rho).}
\usage{
admm_oneclust(
  iclust = 1,
  niter,
  y,
  Dlm1,
  Dl,
  l = NULL,
  TT,
  N,
  dimdat,
  maxdev,
  rho,
  rhoinit = rho,
  Xinv,
  schurA,
  schurB,
  sigmainv,
  lambda,
  resp,
  ylist,
  err_rel = 0.001,
  err_abs = 0,
  zerothresh,
  z,
  w,
  uw,
  uz,
  warmstart = FALSE,
  mu.warm = if (!warmstart) NULL,
  first_iter,
  outer_iter,
  local_adapt,
  sigma,
  sigma_eig_by_clust,
  space = 20,
  norms = F,
  objective = F
)
}
\description{
One cluster's admm for a fixed step size (rho).
}