% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{calc_max_lambda}
\alias{calc_max_lambda}
\title{Estimate maximum lambda values numerically.  First starts with a large
initial value \code{max_lambda_mean} and \code{max_lambda_prob}, and runs
the EM algorithm on decreasing set of values (sequentially halved). This
stops once you see non-simple probabilities or means, and returns the \emph{smallest}
regularization (lambda) value pair that gives full sparsity.}
\usage{
calc_max_lambda(
  ylist,
  countslist = NULL,
  numclust,
  max_lambda_mean = 4000,
  max_lambda_prob = 1000,
  verbose = FALSE,
  iimax = 16,
  ...
)
}
\arguments{
\item{ylist}{List of responses.}

\item{numclust}{Number of clusters.}

\item{max_lambda_mean}{Defaults to 4000.}

\item{max_lambda_prob}{Defaults to 1000.}

\item{iimax}{Maximum value of x for 2^{-x} factors to try.}

\item{...}{Other arguments to \code{flowtrend_once()}.}
}
\value{
list containing the two maximum values to use.
}
\description{
Note that the \code{zero_stabilize=TRUE} option is used in
\code{flowtrend()}, which basically means the EM algorithm runs only until
the zero pattern stabilizes.
}