% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{check_converge}
\alias{check_converge}
\title{Check convergence of ADMM with a fixed step size (rho).}
\usage{
check_converge(
  mu,
  rho,
  w,
  z,
  w_prev,
  z_prev,
  uw,
  uz,
  Dl,
  tDl,
  err_rel = 1e-04,
  err_abs = 0
)
}
\description{
Check convergence of ADMM with a fixed step size (rho).
}