% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{create_oracle}
\alias{create_oracle}
\title{Reformatting |datobj| to create a flowtrend-like object that contains
"oracle" information of the model that
generates the simulated pseudo-real data.}
\usage{
create_oracle(datobj)
}
\arguments{
\item{datobj}{A data object.}
}
\value{
A flowtrend-like object containing mn, sigma, and prob.
}
\description{
Reformatting |datobj| to create a flowtrend-like object that contains
"oracle" information of the model that
generates the simulated pseudo-real data.
}