% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{cv_aggregate}
\alias{cv_aggregate}
\title{Aggregate CV scores from the results, saved in \code{destin}.}
\usage{
cv_aggregate(destin)
}
\arguments{
\item{destin}{Directory with cross-validation output.}
}
\description{
Aggregate CV scores from the results, saved in \code{destin}.
}