% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{cv_flowtrend}
\alias{cv_flowtrend}
\title{Cross-validation for flowtrend(). Saves results to separate files in
\code{destin}.}
\usage{
cv_flowtrend(
  ylist,
  countslist,
  destin = ".",
  lambda_means,
  lambda_probs,
  l,
  l_prob,
  iimat = NULL,
  maxdev,
  numclust,
  nfold,
  nrestart,
  verbose = FALSE,
  refit = FALSE,
  save_meta = FALSE,
  mc.cores = 1,
  folds = NULL,
  seedtab = NULL,
  ...
)
}
\arguments{
\item{ylist}{Data.}

\item{countslist}{Counts corresponding to multiplicities.}

\item{destin}{Directory where output files are saved.}

\item{lambda_means}{Regularization parameters for means.}

\item{lambda_probs}{Regularization parameters for probs.}

\item{l}{Degree of differencing for the mean trend filtering. l=0 will give
you piecewise constant means; l=1 is piecewise linear, and so forth.}

\item{l_prob}{Degree of differencing for the probability trend filtering.}

\item{maxdev}{Maximum deviation of cluster means across time..}

\item{numclust}{Number of clusters.}

\item{nfold}{Number of cross-validation folds. Defaults to 5.}

\item{nrestart}{Number of repetitions.}

\item{verbose}{Loud or not? EM iteration progress is printed.}

\item{refit}{If TRUE, estimate the model on the full data, for each pair of
regularization parameters.}

\item{save_meta}{If TRUE, save meta data.}

\item{mc.cores}{Use this many CPU cores.}

\item{folds}{Manually provide CV folds (list of time points of data to use
as CV folds). Defaults to NULL.}

\item{...}{Additional arguments to flowtrend().}

\item{blocksize}{Contiguous time blocks from which to form CV time folds.}
}
\value{
No return.
}
\description{
Cross-validation for flowtrend(). Saves results to separate files in
\code{destin}.
}