% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{cv_makebest}
\alias{cv_makebest}
\title{From the results saved in \code{destin}, aggregate all |nrestart| files to retain only the "best" restart, and delete the rest.}
\usage{
cv_makebest(destin)
}
\arguments{
\item{destin}{Directory with cross-validation output.}
}
\description{
All meta information (|nfold|, |cv_gridsize|, |nrestart|, |lambda_means|, |lambda_probs|) comes from \code{meta.Rdata}.
}