% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{cv_summary}
\alias{cv_summary}
\title{Main function for summarizing the cross-validation results.}
\usage{
cv_summary(destin = ".", save = FALSE, filename = "summary.RDS")
}
\arguments{
\item{destin}{Directory where output files are saved.}

\item{save}{If TRUE, save to \code{file.path(destin, filename)}.}

\item{filename}{File name to save to.}
}
\value{
List containing summarized results from cross-validation. Here are
some objects in this list: \code{bestres} is the the overall best model
chosen from the cross-validation; \code{cvscoremat} is a 2d matrix of CV
scores from all pairs of regularization parameters; \code{bestreslist} is a
list of all the best models (out of \code{nrestart} EM replications) from the
each pair of lambda values. If \code{isTRUE(save)}, nothing is returned.
}
\description{
Main function for summarizing the cross-validation results.
}