% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{dt2ylist}
\alias{dt2ylist}
\title{Converting to a list of matrices, \code{ylist}, to input to \code{flowtrend()}.}
\usage{
dt2ylist(dt)
}
\arguments{
\item{dt}{Output from \code{gendat_1d()}.}
}
\value{
List of matrices
}
\description{
Converting to a list of matrices, \code{ylist}, to input to \code{flowtrend()}.
}