% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{flowmeans_each}
\alias{flowmeans_each}
\title{Apply flowmeans in each cytogram.}
\usage{
flowmeans_each(ylist, numclust)
}
\arguments{
\item{ylist}{Data.}

\item{numclust}{Number of clusters.}
}
\description{
Apply flowmeans in each cytogram.
}