% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{flowtrend-package}
\alias{flowtrend}
\alias{flowtrend-package}
\title{flowtrend}
\description{
This package implements the \code{flowtrend} method for automatic gating of flow cytometry data using trend filtering.
}
\author{
\strong{Maintainer}: Sangwon Hyun \email{sangwonh@ucsc.edu}

}
\keyword{internal}