% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{flowtrend}
\alias{flowtrend}
\title{Main function. Repeats the EM algorithm (\code{flowtrend_once()}) with |nrep| restarts (5 by default).}
\usage{
flowtrend(..., nrestart = 5)
}
\arguments{
\item{...}{: arguments for \code{flowtrend_once()}}

\item{nrestart}{: number of random restarts}
}
\description{
Main function. Repeats the EM algorithm (\code{flowtrend_once()}) with |nrep| restarts (5 by default).
}