% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{flowtrend_once}
\alias{flowtrend_once}
\title{Estimate flowtrend model once.}
\usage{
flowtrend_once(
  ylist,
  countslist = NULL,
  x = NULL,
  numclust,
  niter = 1000,
  l,
  l_prob = NULL,
  mn = NULL,
  lambda = 0,
  lambda_prob = NULL,
  verbose = FALSE,
  tol_em = 1e-04,
  maxdev = NULL,
  countslist_overwrite = NULL,
  admm = TRUE,
  admm_err_rel = 0.001,
  admm_err_abs = 1e-04,
  admm_local_adapt = FALSE,
  admm_local_adapt_niter = if (admm_local_adapt) 10 else 1,
  seed = NULL
)
}
\arguments{
\item{ylist}{Data.}

\item{countslist}{Counts corresponding to multiplicities.}

\item{x}{Times, if points are not evenly spaced. Defaults to NULL, in which
case the value becomes \code{1:T}, for the $T==length(ylist)$.}

\item{numclust}{Number of clusters.}

\item{niter}{Maximum number of EM iterations.}

\item{l}{Degree of differencing for the mean trend filtering.}

\item{l_prob}{Degree of differencing for the probability trend filtering.}

\item{mn}{Initial value for cluster means. Defaults to NULL, in which case
initial values are randomly chosen from the data.}

\item{lambda}{Smoothing parameter for means}

\item{lambda_prob}{Smoothing parameter for probabilities}

\item{verbose}{Loud or not? EM iteration progress is printed.}

\item{tol_em}{Relative numerical improvement of the objective value at which
to stop the EM algorithm}

\item{maxdev}{Maximum deviation of cluster means across time..}

\item{admm_local_adapt_niter}{}
}
\value{
List object with flowtrend model estimates.
}
\description{
Estimate flowtrend model once.
}