% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gen_diff_mat}
\alias{gen_diff_mat}
\title{Generating Difference Matrix of Specified Order}
\usage{
gen_diff_mat(n, l, x = NULL)
}
\arguments{
\item{n}{length of vector to be differenced}

\item{l}{order of differencing}

\item{x}{optional. Spacing of input points.}
}
\value{
A n by n-l-1 matrix
}
\description{
Generating Difference Matrix of Specified Order
}