% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gen_tf_mat}
\alias{gen_tf_mat}
\title{A lasso regressor matrix H that can be used to solve an equivalent problem as the trend filtering of the \code{k}'th degree.}
\usage{
gen_tf_mat(n, k, x = NULL)
}
\arguments{
\item{n}{Total number of time points.}

\item{k}{Degree of trend filtering for cluster probabilities. $k=0$ is fused lasso, $k=1$ is linear trend filtering, and so on.}

\item{x}{Time points}
}
\value{
$n$ by $n$ matrix.
}
\description{
A lasso regressor matrix H that can be used to solve an equivalent problem as the trend filtering of the \code{k}'th degree.
}