% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gen_tf_mat_equalspace}
\alias{gen_tf_mat_equalspace}
\title{Creates trendfiltering regression matrix using Lemma 2 of Ryan Tibshirani's
trendfilter paper (2014); works on equally spaced data only.}
\usage{
gen_tf_mat_equalspace(n, k)
}
\arguments{
\item{n}{Number of data points}

\item{k}{Order of trend filter. 0 is fused lasso, and so on.}
}
\value{
n by n matrix.
Connects kk to kk-1.
}
\description{
Creates trendfiltering regression matrix using Lemma 2 of Ryan Tibshirani's
trendfilter paper (2014); works on equally spaced data only.
}
\examples{
ord = 1
H_tf <- gen_tf_mat_equalspace(n = 100, k = ord)
H_tf[,1] * 100
H_tf[,2] * 100
H_tf[,3] * 100
H_tf[,4] * 100
H_tf[,99] * 100
H_tf[,100] * 100
}