% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gendat_1d}
\alias{gendat_1d}
\title{Generates some synthetic 1-dimensional data with three clusters. Returns a
data frame with (1) time, (2) Y (3) mu (4) cluster assignments.}
\usage{
gendat_1d(TT, ntlist, offset = 0, return_model = FALSE, sd3 = NULL)
}
\arguments{
\item{TT}{Number of time points.}

\item{offset}{Defaults to 0. How much to push up cluster 1.}

\item{return_model}{If true, return true cluster means and probabilities
instead of data.}

\item{nt}{Number of particles at each time.}
}
\value{
long matrix with data or model.
}
\description{
Generates some synthetic 1-dimensional data with three clusters. Returns a
data frame with (1) time, (2) Y (3) mu (4) cluster assignments.
}