% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gendat_2d}
\alias{gendat_2d}
\title{Generates some synthetic 2-dimensional data with three clusters.}
\usage{
gendat_2d(TT, ntlist)
}
\arguments{
\item{TT}{Number of time points.}

\item{nt}{Number of particles at each time.}
}
\value{
List containing (1) ylist, (2) mnlist, (3) clusterlist.
}
\description{
Generates some synthetic 2-dimensional data with three clusters.
}