% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_rand_from_table}
\alias{get_rand_from_table}
\title{RAND index.}
\usage{
get_rand_from_table(tab)
}
\arguments{
\item{tab}{A matrix containing  a 2 x 2 table in the first \link{1:2,1:2}
entries; e.g., from make_contingency_table().}
}
\value{
Rand index.
}
\description{
RAND index.
}