% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{gmm_each}
\alias{gmm_each}
\title{Apply a gmm in each 1-dimensional cytogram.}
\usage{
gmm_each(one_y, numclust)
}
\arguments{
\item{one_y}{One-column matrix.}

\item{numclust}{Number of clusters.}
}
\description{
Apply a gmm in each 1-dimensional cytogram.
}