% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{interpolate_mn}
\alias{interpolate_mn}
\title{Do a linear interpolation of the cluster means.}
\usage{
interpolate_mn(x, tt, iclust, mn)
}
\arguments{
\item{x}{Training times.}

\item{tt}{Prediction time.}

\item{iclust}{Cluster number.}

\item{mn}{length(x) by dimdat by numclust matrix.}
}
\value{
A dimdat-length vector.
}
\description{
Do a linear interpolation of the cluster means.
}