% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{interpolate_prob}
\alias{interpolate_prob}
\title{Do a linear interpolation of the cluster means.}
\usage{
interpolate_prob(x, tt, iclust, prob)
}
\arguments{
\item{x}{Training times.}

\item{tt}{Prediction time.}

\item{iclust}{Cluster number.}

\item{prob}{length(x) by numclust array or matrix.}
}
\value{
One probability.
}
\description{
Do a linear interpolation of the cluster means.
}