% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{keep_only_best}
\alias{keep_only_best}
\title{Keeping only the output files for the "best" restart, and deleting the rest.}
\usage{
keep_only_best(destin, iprob, imu, ifold, nrestart, best_irestart)
}
\description{
Keeping only the output files for the "best" restart, and deleting the rest.
}