% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{keep_only_best_refit}
\alias{keep_only_best_refit}
\title{Keeping only the output files (among refit models) for the "best" restart,
and deleting the rest.}
\usage{
keep_only_best_refit(destin, iprob, imu, nrestart, best_irestart)
}
\description{
Keeping only the output files (among refit models) for the "best" restart,
and deleting the rest.
}