% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{load_all_refit_objectives}
\alias{load_all_refit_objectives}
\title{Loading all objectives, with NA's for missing files}
\usage{
load_all_refit_objectives(destin, iprob, imu, ifold, nrestart)
}
\description{
Loading all objectives, with NA's for missing files
}