% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{loglik_tt}
\alias{loglik_tt}
\title{Log likelihood for a single time point's data.}
\usage{
loglik_tt(
  ylist,
  tt,
  mu,
  sigma,
  prob,
  dimdat = NULL,
  countslist = NULL,
  numclust
)
}
\arguments{
\item{ylist}{Data.}

\item{tt}{Time point.}

\item{mu}{Cluster means.}

\item{prob}{Cluster variances.}
}
\description{
Log likelihood for a single time point's data.
}