% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_best_cvscore_filename}
\alias{make_best_cvscore_filename}
\title{Create file name (a string) for cross-validation results.}
\usage{
make_best_cvscore_filename(iprob, imu, ifold)
}
\arguments{
\item{irestart}{}
}
\description{
Create file name (a string) for cross-validation results.
}