% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_contingency_table}
\alias{make_contingency_table}
\title{2 x 2 contigency table from membership vectors.}
\usage{
make_contingency_table(mem1, mem2)
}
\arguments{
\item{mem1}{Membership vector.}

\item{mem2}{Another membership vector.}
}
\value{
A 3x3 matrix containing (1) a 2 x 2 table in the first \link{1:2,1:2}
entries, and (2) row sums and column sums and total sums in the \link{,3}, \link{3,}
entries.
}
\description{
2 x 2 contigency table from membership vectors.
}