% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_cv_folds}
\alias{make_cv_folds}
\title{Define the time folds cross-validation.}
\usage{
make_cv_folds(ylist = NULL, nfold, TT = NULL)
}
\arguments{
\item{nfold}{Number of folds.}
}
\value{
List of fold indices.
}
\description{
Define the time folds cross-validation.
}