% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_cv_folds_in_blocks}
\alias{make_cv_folds_in_blocks}
\title{Define the "blocked" time folds for cross-validation.
This means that contiguous of times will be used to define CV folds.}
\usage{
make_cv_folds_in_blocks(ylist = NULL, nfold, TT = NULL, blocksize)
}
\arguments{
\item{nfold}{Number of folds.}

\item{blocksize}{Size of block (e.g. 3 will produce the example above).}
}
\value{
List of fold indices.
}
\description{
The first fold will be ( 1 2 3 16 17 18 31 32 33 46 47 48 61 62 63 76 77 78
91 92 93), the second fold will be (4 5 6 19 20 21 34 35 36 49 50 51 64 65 66
79 80 81 94 95 96), and so forth.
}