% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_iilist}
\alias{make_iilist}
\title{Helper to divide up the jobs in \code{iimat} into a total of
\code{arraynum_max} jobs. The purpose is to divide the jobs, in order to run
this on a server.}
\usage{
make_iilist(arraynum_max, iimat)
}
\arguments{
\item{arraynum_max}{Maximum SLURM array number.}

\item{iimat}{matrix whose rows contain CV job indices.}
}
\description{
Helper to divide up the jobs in \code{iimat} into a total of
\code{arraynum_max} jobs. The purpose is to divide the jobs, in order to run
this on a server.
}