% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_iimat}
\alias{make_iimat}
\title{Indices for the cross validation jobs.}
\usage{
make_iimat(cv_gridsize, nfold, nrestart)
}
\arguments{
\item{cv_gridsize}{CV grid size.}

\item{nfold}{Number of CV folds.}

\item{nrestart}{Number of random restarts of EM algorithm.}
}
\value{
Integer matrix.
}
\description{
The resulting iimat looks like this:
}
\details{
ind    iprob  imu ifold irestart
55      6     1     2    1
56      7     1     2    1
57      1     2     2    1
58      2     2     2    1
59      3     2     2    1
60      4     2     2    1
}