% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_iimat_small}
\alias{make_iimat_small}
\title{2d indices for the cross validation jobs.}
\usage{
make_iimat_small(cv_gridsize, nrestart)
}
\arguments{
\item{cv_gridsize}{CV grid size.}

\item{nrestart}{Number of random restarts of EM algorithm.}
}
\value{
Integer matrix.
}
\description{
The resulting iimat looks like this:
(#, iprob, imu, irestart)
1, 1, 1, 1
2, 1, 2, 1
3, 1, 3, 1
}