% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{make_refit_filename}
\alias{make_refit_filename}
\title{Create file name (a string) for re-estimated models for the lambda values
indexed by \code{iprob} and \code{imu}.}
\usage{
make_refit_filename(iprob, imu, irestart)
}
\arguments{
\item{irestart}{}
}
\description{
Create file name (a string) for re-estimated models for the lambda values
indexed by \code{iprob} and \code{imu}.
}