% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{match_clusters}
\alias{match_clusters}
\title{Apply the hungarian matching algorithm to each time point.}
\usage{
match_clusters(fmns_obj)
}
\arguments{
\item{flowmeans_obj}{}
}
\description{
Apply the hungarian matching algorithm to each time point.
}