% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{match_clusters_gmm}
\alias{match_clusters_gmm}
\title{Sequentially permute a time series of GMMs.}
\usage{
match_clusters_gmm(gmm_list, numclust = 2)
}
\arguments{
\item{gmm_list}{Output of an lapply of gmm_each over ylist.}

\item{numclust}{Number of clusters.}
}
\value{
Same format as |gmm_list|
}
\description{
Sequentially permute a time series of GMMs.
}