% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{matrix_function_solve_triangular_sylvester_barebonesC2}
\alias{matrix_function_solve_triangular_sylvester_barebonesC2}
\title{Solve "barebones" sylvester equation that takes upper triangular matrices as coefficients.}
\usage{
matrix_function_solve_triangular_sylvester_barebonesC2(TA, TB, C)
}
\arguments{
\item{TA}{Upper-triangular matrix}

\item{TB}{Upper-triangular matrix}

\item{C}{matrix}
}
\description{
Solve "barebones" sylvester equation that takes upper triangular matrices as coefficients.
}