% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{memlist_to_respmat}
\alias{memlist_to_respmat}
\title{Convert list of memberships into a list of responsibilities.}
\usage{
memlist_to_respmat(memlist)
}
\arguments{
\item{memlist}{List of memberships}
}
\description{
Convert list of memberships into a list of responsibilities.
}