% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_sigma}
\alias{Mstep_sigma}
\title{M step for cluster covariance (sigma).}
\usage{
Mstep_sigma(resp, ylist, mn, numclust)
}
\arguments{
\item{resp}{Responsibility.}

\item{ylist}{Data.}

\item{mn}{Means}

\item{numclust}{Number of clusters.}
}
\value{
(K x d x d) array containing K (d x d) covariance matrices.
}
\description{
M step for cluster covariance (sigma).
}