% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{my_symmetric_kl}
\alias{my_symmetric_kl}
\title{Given two K x 2 columns with parameters, make K x K distance matrix.}
\usage{
my_symmetric_kl(c1, c2)
}
\arguments{
\item{c1}{Data frame with K rows and 2 columns (mu and sigma)}

\item{c2}{Another data frame of the same size as \code{c1}.}
}
\value{
K x K matrix.
}
\description{
Given two K x 2 columns with parameters, make K x K distance matrix.
}