% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{objective}
\alias{objective}
\title{Evaluating the penalized negative log-likelihood on all data \code{ylist} given parameters \code{mu}, \code{prob}, and \code{sigma}.}
\usage{
objective(
  mu,
  prob,
  prob_link = NULL,
  sigma,
  ylist,
  Dlp1 = NULL,
  l = NULL,
  lambda = 0,
  l_prob = NULL,
  Dlp1_prob = NULL,
  lambda_prob = 0,
  alpha = NULL,
  beta = NULL,
  denslist_by_clust = NULL,
  countslist = NULL,
  unpenalized = FALSE
)
}
\arguments{
\item{unpenalized}{if TRUE, return the unpenalized out-of-sample fit.}
}
\description{
Evaluating the penalized negative log-likelihood on all data \code{ylist} given parameters \code{mu}, \code{prob}, and \code{sigma}.
}