% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{objective_per_cluster}
\alias{objective_per_cluster}
\title{computes the ADMM objective for one cluster}
\usage{
objective_per_cluster(
  y,
  TT,
  d,
  l,
  mu,
  resp,
  Sigma_inv,
  Dlp1,
  Dl,
  maxdev,
  lambda,
  rho,
  N
)
}
\description{
computes the ADMM objective for one cluster
}