% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{one_job}
\alias{one_job}
\title{Helper function to run ONE job for CV, in iprob, imu, ifold, irestart.}
\usage{
one_job(
  iprob,
  imu,
  ifold,
  irestart,
  folds,
  destin,
  lambda_means,
  lambda_probs,
  seedtab = NULL,
  x = NULL,
  ylist,
  countslist,
  l,
  l_prob,
  ...
)
}
\arguments{
\item{iprob}{Index for prob.}

\item{imu}{Index for beta.}

\item{ifold}{Index for CV folds.}

\item{irestart}{Index for 1 through nrestart.}

\item{folds}{CV folds (from \code{make_cv_folds()}).}

\item{destin}{Destination directory.}

\item{lambda_means}{List of regularization parameters for mean model.}

\item{lambda_probs}{List of regularization parameters for prob model.}

\item{ylist}{Data.}

\item{countslist}{Counts or biomass.}

\item{...}{Rest of arguments for \code{flowtrend_once()}.}
}
\value{
Nothing is returned. Instead, a file named "1-1-1-1-cvscore.Rdata"
is saved in \code{destin}. (The indices here are iprob-imu-ifold-irestart).
}
\description{
Helper function to run ONE job for CV, in iprob, imu, ifold, irestart.
}