% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{one_job_refit}
\alias{one_job_refit}
\title{Refit model for one pair of regularization parameter values. Saves to
\code{nrestart} files named like "1-4-3-fit.Rdata", for
"(iprob)-(imu)-(irestart)-fit.Rdata".}
\usage{
one_job_refit(
  iprob,
  imu,
  destin,
  lambda_means,
  lambda_probs,
  l,
  l_prob,
  seedtab = NULL,
  ylist,
  countslist,
  x,
  ...
)
}
\arguments{
\item{iprob}{Index for prob.}

\item{imu}{Index for beta.}

\item{destin}{Destination directory.}

\item{lambda_means}{List of regularization parameters for mean model.}

\item{lambda_probs}{List of regularization parameters for prob model.}

\item{ylist}{Data.}

\item{countslist}{Counts or biomass.}

\item{...}{Rest of arguments for \code{flowtrend_once()}.}
}
\description{
(Note, \code{nrestart} is not an input to this function.)
}