% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{one_symmetric_kl}
\alias{one_symmetric_kl}
\title{Symmetric KL divergence between two Gaussian distributions.}
\usage{
one_symmetric_kl(mu1, mu2, sd1, sd2)
}
\arguments{
\item{mu1}{Mean for distribution 1.}

\item{mu2}{Mean for distribution 2.}

\item{sd1}{Standard deviation for distribution 1.}

\item{sd2}{Standard deviation for distribution 2.}
}
\description{
Symmetric KL divergence between two Gaussian distributions.
}