% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{outer_converge}
\alias{outer_converge}
\title{LA-ADMM requires an convergence check for the outer layer.}
\usage{
outer_converge(objectives)
}
\arguments{
\item{objectives}{Objectives of the outer layer.}
}
\value{
TRUE if relative improvement is smaller than 1E-5 in the last four
outer iterations' objective.
}
\description{
LA-ADMM requires an convergence check for the outer layer.
}