% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{overfit_flowmeans}
\alias{overfit_flowmeans}
\title{Apply flowmeans sequentially and then cluster match}
\usage{
overfit_flowmeans(ylist, numclust, verbose = FALSE)
}
\arguments{
\item{ylist}{Data.}

\item{numclust}{Number of clusters.}
}
\description{
Apply flowmeans sequentially and then cluster match
}