% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{overfit_gmm}
\alias{overfit_gmm}
\title{Pool the entire series of cytograms, fit a GMM model, and re-aggregate parameters.}
\usage{
overfit_gmm(ylist, numclust = 2, reorder = TRUE)
}
\arguments{
\item{ylist}{Data.}

\item{numclust}{Number of clusters (only works for 2).}
}
\description{
Pool the entire series of cytograms, fit a GMM model, and re-aggregate parameters.
}