% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{parse_args}
\alias{parse_args}
\title{Parse command line arguments and assigns the values of them. |args| is meant
to just be additional command line arguments.}
\usage{
parse_args(args, verbose = FALSE)
}
\arguments{
\item{args}{Argument.}
}
\value{
Nothing.
}
\description{
Parse command line arguments and assigns the values of them. |args| is meant
to just be additional command line arguments.
}