% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d}
\alias{plot_1d}
\title{Makes 1d plot of data and model}
\usage{
plot_1d(ylist, countslist = NULL, obj = NULL, x = NULL, add_point = FALSE)
}
\arguments{
\item{ylist}{Data.}

\item{obj}{flowtrend object. Defaults to NULL.}

\item{x}{time points. Defaults to NULL.}
}
\value{
ggplot object with data, and optionally, a flowtrend model overlaid.
}
\description{
Makes 1d plot of data and model
}