% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d_add_model}
\alias{plot_1d_add_model}
\title{Add a model.}
\usage{
plot_1d_add_model(gg0, obj, idim, plot_band = TRUE)
}
\arguments{
\item{gg0}{result of plot_1d}

\item{obj}{Model}

\item{idim}{dimension}

\item{plot_band}{If TRUE, plot a +1.96 standard deviation band.}
}
\value{
ggplot object
}
\description{
Add a model.
}