% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d_flowmeans}
\alias{plot_1d_flowmeans}
\title{Plotter for underfit and overfit flowmeans.}
\usage{
plot_1d_flowmeans(obj, ylist, countslist = NULL, bin = FALSE)
}
\arguments{
\item{obj}{Output from \code{underfit_flowmeans()}}

\item{ylist}{Data}

\item{countslist}{Optional: counts for each ylist}

\item{bin}{Whether data is binned.}
}
\value{
ggplot object
}
\description{
Plotter for underfit and overfit flowmeans.
}