% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d_with_membership}
\alias{plot_1d_with_membership}
\title{Makes 1d plot of data when}
\usage{
plot_1d_with_membership(
  ylist,
  memlist,
  countslist = NULL,
  x = NULL,
  alpha = 0.01
)
}
\arguments{
\item{ylist}{Particlel-level data. A list of (|nt| by |dimdat|) matrices.}

\item{memlist}{List of memberships for the particles.}

\item{x}{Time points. Defaults to NULL.}

\item{alpha}{Between 0 and 1, how transparent to plot the data
points. Defaults to 0.1.}
}
\value{
ggplot object with data, and optionally, a flowtrend model overlaid.
}
\description{
Makes 1d plot of data when
}