% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_from_summary}
\alias{plot_from_summary}
\title{Makes a series of plots.}
\usage{
plot_from_summary(destin)
}
\arguments{
\item{destin}{This directly has (1) summary.RDS, (2) meta.Rdata, and (3) datobj.RDS; optionally, it may have a dt_model}
}
\value{
Nothing; but the pdf plots are created.
}
\description{
Makes a series of plots.
}