% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_prob}
\alias{plot_prob}
\title{Makes cluster probability plot (lines over time).}
\usage{
plot_prob(obj, x = NULL)
}
\arguments{
\item{obj}{Estimated model (from e.g. \code{flowtrend()})}
}
\description{
Makes cluster probability plot (lines over time).
}