% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{predict_flowtrend}
\alias{predict_flowtrend}
\title{Prediction: Given new timepoints in the original time interval,generate a set
of means and probs (and return the same Sigma).}
\usage{
predict_flowtrend(obj, newtimes = NULL)
}
\arguments{
\item{obj}{Object returned from covariate EM flowtrend().}

\item{newtimes}{New times at which to make predictions.}
}
\value{
List containing mean, prob, and sigma, and x.
}
\description{
Prediction: Given new timepoints in the original time interval,generate a set
of means and probs (and return the same Sigma).
}