% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{projCmat}
\alias{projCmat}
\title{Project /rows/ of matrix |mat| into a ball of size C.}
\usage{
projCmat(mat, C)
}
\arguments{
\item{mat}{Matrix whose rows will be projected into a C-sized ball.}

\item{C}{radius}
}
\value{
Projected matrix.
}
\description{
Project /rows/ of matrix |mat| into a ball of size C.
}