% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{prox_dp}
\alias{prox_dp}
\title{Implements the dynamic programming algorithm for the 1D Fused Lasso proximal operator.}
\usage{
prox_dp(z, lam)
}
\arguments{
\item{z}{A numeric vector of input values ($\mathbf{y}$).}

\item{lam}{A numeric scalar, the regularization parameter ($\lambda$).}
}
\value{
A numeric vector, the result of the proximal operation ($\boldsymbol{\beta}$).
}
\description{
Implements the dynamic programming algorithm for the 1D Fused Lasso proximal operator.
}