% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{rand_from_mems}
\alias{rand_from_mems}
\title{Faster rand index calculation from membership vectors.}
\usage{
rand_from_mems(mem1, mem2)
}
\arguments{
\item{mem2}{}
}
\value{
RAND index.
}
\description{
Faster rand index calculation from membership vectors.
}