% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{rand_old}
\alias{rand_old}
\title{A "soft" version of a rand index between two sets of responsibility
(membership probability) matrices. Measures the /disagreement/ between the two clusterings.}
\usage{
rand_old(resp_list1, resp_list2, times = NULL, prop = 0.25)
}
\arguments{
\item{resp_list1}{One list of responsibility matrices.}

\item{resp_list2}{Another list of responsibility matrices.}

\item{times}{Optional; if you would like to isolate your attention to some specific times.}
}
\value{
A single soft rand index number
}
\description{
A "soft" version of a rand index between two sets of responsibility
(membership probability) matrices. Measures the /disagreement/ between the two clusterings.
}