% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{reorder_gmm_fit}
\alias{reorder_gmm_fit}
\title{Reordering function for gmm results.}
\usage{
reorder_gmm_fit(obj, new_order)
}
\arguments{
\item{obj}{result of \code{gmm_overfit()} or \code{gmm_underfit()}.}

\item{new_order}{new ordering of clusters to use.}
}
\value{
reordered object.
}
\description{
Reordering function for gmm results.
}