% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{reorder_kl}
\alias{reorder_kl}
\title{Reorder the cluster numbers for a new flowtrend object \code{newres}; the best
permutation (reordering) is to match the original flowmix object
\code{origres}.}
\usage{
reorder_kl(newres, origres, ylist_particle, fac = 100, verbose = FALSE)
}
\arguments{
\item{newres}{New flowtrend object to reorder.}

\item{origres}{Original flowtrend object.}

\item{ylist_particle}{The particle-level data.}

\item{fac}{Defaults to 100, to take 1/100'th of the particles from each time point.}

\item{verbose}{Loud or not?}
}
\value{
Reordered res
}
\description{
Reorder the cluster numbers for a new flowtrend object \code{newres}; the best
permutation (reordering) is to match the original flowmix object
\code{origres}.
}