% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{soft_gate_one_responsibility_matrix}
\alias{soft_gate_one_responsibility_matrix}
\title{Soft-gates a responsibility matrix by a bernoulli (or multinoulli) draw.}
\usage{
soft_gate_one_responsibility_matrix(oneresp)
}
\arguments{
\item{oneresp}{A 2-column responsibility matrix}
}
\description{
Soft-gates a responsibility matrix by a bernoulli (or multinoulli) draw.
}