% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{softmax}
\alias{softmax}
\title{Softmax function.}
\usage{
softmax(prob_link)
}
\arguments{
\item{prob_link}{alpha, which is a (T x K) matrix.}
}
\value{
exp(alpha)/rowSum(exp(alpha)). A (T x K) matrix.
}
\description{
Softmax function.
}