% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{symmetric_kl_between_gaussians}
\alias{symmetric_kl_between_gaussians}
\title{Between two sets of Gaussian mean/sd parameters, what is the symmetric KL
distance.}
\usage{
symmetric_kl_between_gaussians(param1, param2, numclust = 2)
}
\arguments{
\item{param1}{list of mn1, mn2, sd1, sd2}

\item{param2}{another such list.}
}
\value{
(numclust x numclust) distance matrix.
}
\description{
Between two sets of Gaussian mean/sd parameters, what is the symmetric KL
distance.
}