% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{underfit_gmm}
\alias{underfit_gmm}
\title{Pool the entire series of cytograms, fit a GMM model, and re-aggregate parameters.}
\usage{
underfit_gmm(ylist, numclust)
}
\arguments{
\item{ylist}{Data.}

\item{numclust}{Number of clusters.}
}
\description{
Pool the entire series of cytograms, fit a GMM model, and re-aggregate parameters.
}