% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{W_update_fused}
\alias{W_update_fused}
\title{Solve a fused lasso problem for the W update. Internally, a fused lasso dynamic
programming solver \code{prox()} (which calls \code{prox_dp()} written in C)
is used.}
\usage{
W_update_fused(l, TT, mu, uw, rho, lambda)
}
\description{
Solve a fused lasso problem for the W update. Internally, a fused lasso dynamic
programming solver \code{prox()} (which calls \code{prox_dp()} written in C)
is used.
}