% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{admm_oneclust}
\alias{admm_oneclust}
\title{One cluster's admm for a fixed step size (rho).}
\usage{
admm_oneclust(
  iclust = 1,
  niter,
  y,
  Dl,
  tDl,
  Dlp1,
  l = NULL,
  TT,
  N,
  dimdat,
  maxdev,
  rho,
  rhoinit = rho,
  Xinv,
  schurA,
  schurB,
  sigmainv,
  lambda,
  resp,
  resp_sum,
  ylist,
  err_rel = 0.001,
  err_abs = 0,
  zerothresh,
  mu,
  z,
  w,
  uw,
  uz,
  first_iter,
  outer_iter,
  local_adapt,
  sigma,
  sigma_eig_by_clust
)
}
\description{
One cluster's admm for a fixed step size (rho).
}