% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{admm_oneclust}
\alias{admm_oneclust}
\title{One cluster's ADMM for a fixed step size (\code{rho}).}
\usage{
admm_oneclust(
  iclust = 1,
  niter,
  y,
  Dl,
  tDl,
  Dlp1,
  l = NULL,
  TT,
  N,
  dimdat,
  maxdev,
  rho,
  rhoinit = rho,
  Xinv,
  schurA,
  schurB,
  sigmainv,
  lambda,
  resp,
  resp_sum,
  ylist,
  err_rel = 0.001,
  err_abs = 0,
  zerothresh,
  mu,
  z,
  w,
  uw,
  uz,
  first_iter,
  em_iter,
  outer_iter,
  local_adapt,
  sigma,
  sigma_eig_by_clust
)
}
\arguments{
\item{iclust}{}

\item{niter}{}

\item{y}{}

\item{Dl}{}

\item{tDl}{}

\item{Dlp1}{}

\item{l}{Default: NULL}

\item{TT}{}

\item{N}{}

\item{dimdat}{}

\item{maxdev}{}

\item{rho}{}

\item{rhoinit}{Default: rho}

\item{Xinv}{}

\item{schurA}{}

\item{schurB}{}

\item{sigmainv}{}

\item{lambda}{}

\item{resp}{}

\item{resp_sum}{}

\item{ylist}{}

\item{err_rel}{Default: 1e-3}

\item{err_abs}{Default: 0}

\item{zerothresh}{}

\item{mu}{}

\item{z}{}

\item{w}{}

\item{uw}{}

\item{uz}{}

\item{first_iter}{}

\item{em_iter}{}

\item{outer_iter}{}

\item{local_adapt}{}

\item{sigma}{}

\item{sigma_eig_by_clust}{}
}
\value{
A list containing the updated parameters for the cluster.
}
\description{
One cluster's ADMM for a fixed step size (\code{rho}).
}
\seealso{
\code{\link{la_admm_oneclust}} which is a wrapper over this function.
}