% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{aug_lagr}
\alias{aug_lagr}
\title{computes the Augmented lagrangian.}
\usage{
aug_lagr(
  y,
  TT,
  d,
  z,
  w,
  l,
  uz,
  uw,
  mu,
  resp,
  Sigma_inv,
  Dlp1,
  Dl,
  maxdev,
  lambda,
  rho,
  N
)
}
\description{
computes the Augmented lagrangian.
}