% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Estep}
\alias{Estep}
\title{E step, which updates the "responsibilities", which are posterior membership probabilities of each particle.}
\usage{
Estep(
  mn,
  sigma,
  prob,
  ylist = NULL,
  numclust,
  denslist_by_clust = NULL,
  first_iter = FALSE,
  countslist = NULL,
  padding = 1e-20
)
}
\arguments{
\item{sigma}{covariance}

\item{padding}{A small amount of padding to add to the weighted
densities. Note, a very small value (like 1E-20) should be used, otherwise
the probabilitistic (soft) gating can be affected quite noticably.}
}
\description{
E step, which updates the "responsibilities", which are posterior membership probabilities of each particle.
}