% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_mu}
\alias{Mstep_mu}
\title{Computes the M step for mu. TODO: use templates for the argument. As shown
here:
https://stackoverflow.com/questions/15100129/using-roxygen2-template-tags}
\usage{
Mstep_mu(
  resp,
  ylist,
  lambda = 0.5,
  l = 3,
  sigma,
  sigma_eig_by_clust = NULL,
  Dlsqrd,
  Dl,
  Dlp1,
  TT,
  N,
  dimdat,
  first_iter = TRUE,
  e_mat,
  mus = NULL,
  Zs = NULL,
  Ws = NULL,
  uws = NULL,
  uzs = NULL,
  maxdev = NULL,
  x = NULL,
  niter = (if (local_adapt) 100 else 10000),
  err_rel = 0.001,
  err_abs = 0,
  zerothresh = 1e-06,
  local_adapt = FALSE,
  local_adapt_niter = 10,
  space = 50
)
}
\arguments{
\item{resp}{Responsbilities of each particle.}

\item{ylist}{}

\item{space}{}
}
\description{
Computes the M step for mu. TODO: use templates for the argument. As shown
here:
https://stackoverflow.com/questions/15100129/using-roxygen2-template-tags
}