% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_mu}
\alias{Mstep_mu}
\title{Computes the M step for mu.}
\usage{
Mstep_mu(
  resp,
  ylist,
  lambda = 0.5,
  l = 3,
  sigma,
  sigma_eig_by_clust = NULL,
  Dlsqrd,
  Dl,
  tDl,
  Dlp1,
  TT,
  N,
  dimdat,
  first_iter = TRUE,
  e_mat,
  mus = NULL,
  Zs = NULL,
  Ws = NULL,
  uws = NULL,
  uzs = NULL,
  maxdev = NULL,
  x = NULL,
  niter = (if (local_adapt) 100 else 1000),
  err_rel = 0.001,
  err_abs = 0,
  zerothresh = 1e-06,
  local_adapt = FALSE,
  local_adapt_niter = 10,
  rho_init = 0.01,
  iter = NULL
)
}
\arguments{
\item{resp}{Responsibilities of each particle.}

\item{ylist}{}

\item{iter}{}
}
\description{
Computes the M step for mu.
}