% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_mu_cvxr}
\alias{Mstep_mu_cvxr}
\title{Testing against \code{Mstep_mu()}, for ONE cluster.}
\usage{
Mstep_mu_cvxr(
  ylist,
  resp,
  lambda,
  l,
  Sigma_inv,
  thresh = 1e-08,
  maxdev = NULL,
  dimdat,
  N,
  ecos_thresh = 1e-08,
  scs_eps = 1e-05
)
}
\arguments{
\item{Sigma_inv}{inverse of Sigma}
}
\description{
Testing against \code{Mstep_mu()}, for ONE cluster.
}