% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{Mstep_mu_cvxr}
\alias{Mstep_mu_cvxr}
\title{Used as a fallback or to test against \code{Mstep_mu()}.}
\usage{
Mstep_mu_cvxr(
  ylist,
  resp,
  lambda,
  l,
  Sigma_inv,
  x = NULL,
  thresh = 1e-08,
  maxdev = NULL,
  dimdat,
  N,
  ecos_thresh = 1e-08,
  scs_eps = 1e-05
)
}
\arguments{
\item{Sigma_inv}{inverse of Sigma}

\item{x}{covariates}
}
\description{
Used as a fallback or to test against \code{Mstep_mu()}.
}