% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_1d}
\alias{plot_1d}
\title{Makes 1d plot of data and model}
\usage{
plot_1d(
  ylist,
  countslist = NULL,
  obj = NULL,
  x = NULL,
  idim = NULL,
  alpha = 0.1,
  bin = FALSE
)
}
\arguments{
\item{ylist}{Data. A list of (|nt| by |dimdat|) matrices}

\item{obj}{A flowtrend (or flowmix) object. Defaults to NULL.}

\item{x}{Time points. Defaults to NULL.}

\item{idim}{If provided, plots the |idim|'th dimension of the data (only particles).}

\item{alpha}{Optional; between 0 and 1, how transparent to plot the data
points.}

\item{bin}{If TRUE, the data is binned.}
}
\value{
ggplot object with data, and optionally, a flowtrend model overlaid.
}
\description{
Makes 1d plot of data and model
}