% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_2d}
\alias{plot_2d}
\title{Simple plotter for 2d particle data.}
\usage{
plot_2d(
  ylist,
  countslist = NULL,
  obj = NULL,
  tt,
  bin = TRUE,
  point_color = "blue",
  raster_colours = c("white", "blue"),
  zero_one_list_censored = NULL
)
}
\arguments{
\item{ylist}{Data. A list of (|nt| by |dimdat|) matrices}

\item{countslist}{Count data.}

\item{obj}{flowtrend (or flowmix) object.}

\item{zero_one_list_censored}{A list of zeros and 1s marking the censored
points.}

\item{time}{Out of 1 through \code{lengthy(list)}, which time point to plot.}
}
\value{
ggplot object.
}
\description{
Simple plotter for 2d particle data.
}