% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_3d}
\alias{plot_3d}
\title{Makes three 2-dimensional plots of the 3d data}
\usage{
plot_3d(
  ylist,
  obj = NULL,
  tt,
  countslist = NULL,
  mn_colours = NULL,
  labels = NULL,
  bin = TRUE,
  plot_title = NULL,
  return_list_of_plots = FALSE,
  zero_one_list_censored = NULL
)
}
\arguments{
\item{ylist}{Data. A list of (|nt| by |dimdat|) matrices}

\item{obj}{flowmix or flowtrend object}

\item{tt}{Time point}

\item{countslist}{Count data.}

\item{return_list_of_plots}{If TRUE, return the list of three plots instead
of the combined plot}

\item{zero_one_list_censored}{A list of zeros and 1s marking the censored
points.}
}
\description{
Makes three 2-dimensional plots of the 3d data
}